+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 * 2025-07-11: Minor changes regarding interface
+* 2025-09-26: Simplified pipeline. Fields are now directly calculated using the voltage wave ratio.